feat: improve code structure!
